Initial commit
The first commit of the server.
The goal of this project is to create a online server that allows a user to interact with the devices they have setup in their homes. With the use of an ID, they can send and receive data from and to the devices of their home. For now, the server is setup using basic express code and I'm hoping to further expand this into a fully fledged system that will control my home.