Added Socket Io functionality
The socket io functionality works now. It will help have a persistent connection between the devices and the server so the devices don't have to continously perform get requests from the server. It sucks that the server now requires a html document to run properly, but maybe we can circumvent this in the near future.