Deploy to web
The application is far from over. But it would be a good idea to have a barebones
version of the application running on the web to test out how to the devices connect to
the server. So this commit is basically a saving point before the application is deployed to heroku.