Change of architecture (Removing Socket IO)
ESP32 doesn't appear to integrate into Socket IO very well.
After searching for a library that allows socket io to work well with
the main device in this network, the ESP32, it seems that the only way
to communicate effectively is to repeatedly ping the server for the information
it needs.