Init MQTT Functionality
The project is now moving away from yet another communication protocol.
The previous implementations worked well with the server, but didn't
communicate as well with the other devices. This communication protocol
is lightweight and robust. Hence why the server will now be using MQTT
to communicate with the rest of the NioT network.