Added cardInfo, cors, morgan and changed palette
To help with server maintainence, morgan was added to monitor the
requests made to the server. For development, cors policy was excluded.
Now the front-end will receive the card information for it's dashboard
through the cardInfo route, enabling a more versatile dashboard
information display. The cardInfo will still need to be further enhanced
to obtain it's information from a Database. The palette route was
simplified and changing the palette will send back the active palette
(Basically echo from the user input). The front-end and back-end are now
working with each other and the network is now able to change the
palette color from the front-end.