Adding dockerfile and docker integration
The application now has some integration with docker conntainers. The
container itself is going to be used to containerize the application and
have the server and the client run together in multiple containers
instead of them running as their own processes in a host machine.
Hopefully this application can then run in any machine that the
application is available in.